// frontend/api/generated.ts
export interface paths {
  '/v1/tours': {
    parameters: {
      query?: never
      header?: never
      path?: never
      cookie?: never
    }
    /**
     * 全ツアー一覧を取得
     * @description すべてのツアー情報を目的地とエリア情報と共に取得します
     */
    get: {
      parameters: {
        query?: never
        header?: never
        path?: never
        cookie?: never
      }
      requestBody?: never
      responses: {
        /** @description ツアー一覧の取得に成功 */
        200: {
          headers: {
            [name: string]: unknown
          }
          content: {
            'application/json': {
              data: {
                tour: {
                  id: number
                  title: string
                  minPriceTaxIncluded: number
                  departsAirportId: number
                  days: number
                  isDirectFlight: boolean
                  airlinesId: number
                  hotelId: number
                  thumbnailFileName: string
                }
                destination: {
                  id: number
                  name: string
                  nameJp: string
                  imageFilename: string
                }
                area: {
                  name: string
                  nameJp: string
                }
              }[]
            }
          }
        }
        /** @description サーバーエラー */
        500: {
          headers: {
            [name: string]: unknown
          }
          content: {
            'application/json': {
              error: string
            }
          }
        }
      }
    }
    put?: never
    post?: never
    delete?: never
    options?: never
    head?: never
    patch?: never
    trace?: never
  }
// ...
}